+++ src/ClosedXML.Parser.Ast/StructureReferenceNode.cs
@@ -1,5 +1,3 @@
-﻿using System.Text;
-
 namespace ClosedXML.Parser;
 
 public record StructureReferenceNode(
@@ -10,24 +8,6 @@ public record StructureReferenceNode(
 {
     public override string GetDisplayString(ReferenceStyle style)
     {
-        var sb = new StringBuilder();
-        if (Table is not null)
-            sb.Append(Table);
-
-        sb.Append('[');
-
-        var list = new List<string>(3);
-        if (Area != StructuredReferenceArea.None)
-            list.Add(Area.GetDisplayString());
-
-        if (FirstColumn is not null)
-            list.Add($"[{FirstColumn}]");
-
-        if (LastColumn is not null)
-            list.Add($"[{LastColumn}]");
-
-        sb.Append(string.Join(",", list));
-        sb.Append(']');
-        return sb.ToString();
+        return $"{Table}{StructuredReferenceWriter.Specifier(Area, FirstColumn, LastColumn)}";
     }
 };